+++ vite/test/types/camel/react-email-rails.d.ts
@@ -1,0 +1,7 @@
+import "react-email-rails"
+
+declare module "react-email-rails" {
+  interface Register {
+    propKeys: "camel"
+  }
+}